refactor extension service
